+++ CHANGELOG.md
@@ -5,10 +5,11 @@
 ### Added
 
 - Parallelization in AST and IR parsing
+- Location resolution for EraVM-specific messages in standard JSON output
 
 ### Changed
 
-- All errors are now written to JSON output in standard JSON mode
+- All messages are now written to JSON output in standard JSON mode
 - AST and IRs are not emitted anymore if not explicitly requested
 - Empty files and contracts are pruned from standard JSON output
 - ASCII boxes are made more compatible with original messages of solc
@@ -19,7 +20,6 @@
 - Dependency graph inefficiency that caused excessive compilation time
 - Removed JSON stream readers which are much slower than strings and vectors
 - Filtered out several EVM-specific warnings from standard JSON output
-- File path resolution for EraVM-specific warnings in standard JSON output
 - Missing output with non-canonical input paths in combined JSON output
 
 ## [1.5.0] - 2024-06-10